% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brands.R
\name{brands}
\alias{brands}
\title{brands}
\usage{
brands(x = "nyx")
}
\arguments{
\item{x}{Character Vector. User inputs desired makeup brand. The input should be wrapped in quotations marks. Set to nyx as default.}
}
\value{
Use this function to get a dataframe for a specific brand.
}
\description{
Obtain dataframe that displays makeup products of a specific brand.
}
\details{
This function allows the user to obtain a dataframe that contains all makeup products of a specific brand the user inputs.
}
\examples{
brands()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{makeup}
