% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_get.R
\name{check_get}
\alias{check_get}
\title{status}
\usage{
check_get()
}
\value{
Use this function to get connection status.
}
\description{
Get the connection status of makeup-api API.
}
\details{
This function allows the user to check the connection with the makeup-api API.
}
\examples{
check_get()
}
\keyword{API}
\keyword{GET,}
\keyword{endpoint,}
