% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorCos.R
\name{colorCos}
\alias{colorCos}
\title{colorCos}
\usage{
colorCos(x = "lipstick", y = "nyx")
}
\arguments{
\item{x}{Character Vector. This argument allows user to input a makeup product type the user is interested in. The string should be wrapped in quotations marks. Set to lipstick as default.}

\item{y}{Character Vector. This argument allows user to input a makeup brand the user is interested in. The string should be wrapped in quotations marks. Set to nyx as default.}
}
\value{
Use this function to get dataframe of colors for a makeup product.
}
\description{
Obtain dataframe of colors of a makeup product from makeup-api API.
}
\details{
This function allows the user to input a makeup product type and makeup brand to obtain a dataframe from which the user would then select the product ID to render another dataframe that shows the corresponding product ID's available colors.
}
\examples{
colorCos()
}
\keyword{API,}
\keyword{color,}
\keyword{endpoint,}
\keyword{lists}
\keyword{makeup,}
