% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmetics.R
\name{cosmetics}
\alias{cosmetics}
\title{cosmetics}
\usage{
cosmetics()
}
\value{
Use this function to get original dataframe.
}
\description{
Get the complete, original dataframe of makeup-api API.
}
\details{
This function allows the user to obtain the oiginal dataframe the makeup-api API. Columns with NA values have been removed.
}
\examples{
cosmetics()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{makeup}
