% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCos.R
\name{interaCos}
\alias{interaCos}
\title{interaCos}
\usage{
interaCos()
}
\value{
Use this function to get an interactive dataframe.
}
\description{
Obtain interactive dataframe of the makeup-api API.
}
\details{
This function generates an interactive makeup-api API dataframe.
}
\examples{
interaCos()
}
\keyword{API}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{interactive,}
