% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookCos.R
\name{lookCos}
\alias{lookCos}
\title{lookCos}
\usage{
lookCos(y = "lipstick", z = "nyx")
}
\arguments{
\item{y}{Character Vector. User inputs desired makeup product type. The input should be wrapped in quotations marks. Set to lipstick as default.}

\item{z}{Character Vector. User inputs desired makeup brand. The string should be wrapped in quotations marks. Set to nyx as default.}
}
\value{
Use this function to get a dataframe taht contains type of makeup product from a brand.
}
\description{
Looks up makeup products for a certin brand.
}
\details{
This function allows the user to obtain a dataframe based on the product type and brand of makeup they want.
}
\examples{
lookCos()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{input}
\keyword{makeup,}
