% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looks.R
\name{looks}
\alias{looks}
\title{looks}
\usage{
looks(smokey = FALSE, natural = FALSE, glam = FALSE)
}
\arguments{
\item{smokey}{Logical Vector. Set to FALSE as default. When set to TRUE, the function will generate a dataframe that contains an eye, lip, a nd face  makeup product that works for a smokey makeup look.}

\item{natural}{Logical Vector. Set to FALSE as default. When set to TRUE, the function will generate a dataframe that contains an eye, lip, a nd face  makeup product that works for a natural makeup look.}

\item{glam}{Logical Vector. Set to FALSE as default. When set to TRUE, the function will generate a dataframe that contains an eye, lip, a nd face  makeup product that works for a glam makeup look.}
}
\value{
Use this function to get connection status.
}
\description{
Obtain dataframe for suggested products to create desired look.
}
\details{
This function allows user to define a look that they want. 
The resulting dataframe will be one that randomly selected products that are suitable for the desired look.
The resulting dataframe will consist of an eye, lip, and foundation makeup product.
}
\examples{
texture()
}
\keyword{API,}
\keyword{character,}
\keyword{dataframe}
\keyword{description,}
\keyword{endpoint,}
\keyword{texture,}
