% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{price}
\alias{price}
\title{price}
\usage{
price(x = 5, y = 50)
}
\arguments{
\item{x}{Numerical  argument. Enter desired lower-bound price point. The resulting dataframe will display products greater than or equal to this numerical value. Default value is 5.}

\item{y}{Numerical  argument. Enter desired upper-bound price point. The resulting dataframe will display products less than or equal to this numerical value. Default value is 50.}
}
\value{
Use this function to get dataframe based on price limits.
}
\description{
Obtain a dataframe based on desired price.
}
\details{
This function allows the user to obtain a dataframe for a specific price range through input of desired price range.
The user will be able to enter the lower-bound and upper-bound price points of their desired price range. The resulting dataframe will be sorted in ascending order.
}
\examples{
price()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{numeric}
\keyword{price,}
