% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCos.R
\name{searchCos}
\alias{searchCos}
\title{searchCos}
\usage{
searchCos(price = TRUE, type = TRUE, brand = TRUE)
}
\arguments{
\item{price}{Logical statement. This is set to TRUE by default. The default setting allows user to
input a price limit for the resulting dataframe. Setting this statement to FALSE would prompt the function to not include price as a condition for the resulting dataframe.}

\item{type}{Logical statement. This is set to TRUE by default. The default setting allows user to input a makeup product type for the resulting dataframe. The resulting dataframe would show only makeup products of this product type. Setting this statement to FALSE would prompt the function to not include makeup product type as a condition for the resulting dataframe.}

\item{brand}{Logical statement. This is set to TRUE by default. The default setting allows user to input a makeup brand for the resulting dataframe. The resulting dataframe would show only makeup products of this makeup brand. Setting this statement to FALSE would prompt the function to not include makeup brand as a condition for the resulting dataframe.}
}
\value{
Use this function to search for makeup products.
}
\description{
Search for desired makeup product from makeup-api API through three conditions.
}
\details{
This function shows the user a dataframe that contains products from three conditions that the user sets.
}
\examples{
searchCos()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{ifelse,}
\keyword{input}
\keyword{logical,}
