% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture.R
\name{texture}
\alias{texture}
\title{texture}
\usage{
texture(x = "matte")
}
\arguments{
\item{x}{Character Vector. Input a texture or trait of a makeup product you are interested in. The string should be wrapped in quotations marks. Set to "matte" as default.}
}
\value{
Use this function to get dataframe based on  trait of makeup product.
}
\description{
Search makeup-api API based on description
}
\details{
This function allows the user to input a desired description they want from a makeup product out of the makeup-api API.
Specifically, this function is designed to allow the user to input a texture of a makeup product such as "matte" or "glossy"
and have the ouput be a dataframe that contains products with that description.
}
\examples{
texture()
}
\keyword{API,}
\keyword{character,}
\keyword{dataframe}
\keyword{description,}
\keyword{endpoint,}
\keyword{texture,}
