% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeCos.R
\name{typeCos}
\alias{typeCos}
\title{typeCos}
\usage{
typeCos(x = "lipstick")
}
\arguments{
\item{x}{Character Vector. User inputs desired makeup product type. This function allows the user to obtain a dataframe that contains all makeup products of a specific product type the user inputs.  The string should be wrapped in quotations marks. Set to lipstick as default.}
}
\value{
Use this function to get a dataframe for a type  of makeup product.
}
\description{
Get the complete, original dataframe of makeup-api API.
}
\details{
This function allows the user to obtain the oiginal dataframe the makeup-api API. Columns with NA values have been removed.
}
\examples{
typeCos()
}
\keyword{API,}
\keyword{dataframe,}
\keyword{endpoint,}
\keyword{input}
\keyword{makeup,}
